requirements.txt
scipy
numpy
annotationframeworkclient
cloud-volume
Pillow
fastremap
tqdm
imageio